frame counter; reduce unnecessary processing
Instead of reading, resizing, converting, and appending every single frame in the video, this code maintains a counter and frame step interval to selectively process only frames that are necessary in (total_frames/frame_step/2) intervals. This reduces the overall processing time and splits the frame selection in relatively equal intervals across the video.