Use regexp to find classname
Sometimes, data can be stored in the same directory, the filename can
give the classname (label) instead of a directory. So, here, we try to
use the classname using a regexp that uses `{classname}` replaced by
`(.*?)`. That's probably not an optimal way, but we need it...

TODO: make research on that topic