Fix problem with frame counter
Some videos can have a really bad header where frame counter is wrong.
In that case, the frame extraction is broken. We can detect if the
lenght of frame is very bad and stop to use header in that case.
That's what is done here. But we probably need to allow XXXGenerator to
not use header at all.